requirements.txt
bleach==3.1.5
requests==2.23.0
python-dotenv==0.13.0
beautifulsoup4==4.9.1
tinydb==4.1.1
Flask==1.1.2
